requirements.txt
discord<=1.0
asyncio
pymysql
requests
pillow
numpy